requirements.txt
fastapi
uvicorn
streamlit
pandas
python-dotenv
google-generativeai
requests
